Added screen,role, screen group and user mapping

+++ auth-service/src/main/java/com/microservice/erp/domain/repositories/RoleRepository.java
@@ -4,10 +4,11 @@ import com.microservice.erp.domain.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, BigInteger> {
     Optional<Role> findRoleByRoleName(String name);
 
     Role findByIsOpenUser(Character isOpenUser);
